+++ CHANGELOG.rst
@@ -5,6 +5,7 @@ Development
 ***********
 
 - Extend explorer to use all implemented APIs
+- Fix cli/restapi: return json and use NULL instead of NaN
 
 0.27.0 (16.02.2022)
 *******************